SED-1192 implemented Vad's feedback
