+++ projects/step-frontend/src/lib/modules/execution/components/execution-item-subtabs/execution-items-subtabs.component.scss
@@ -1,0 +1,6 @@
+.tab-content {
+  display: flex;
+  align-items: center;
+  justify-content: space-around;
+  min-width: 12rem;
+}
