+++ projects/step-frontend/src/lib/modules/execution/execution.module.ts
@@ -15,8 +15,7 @@ import { ReportNodesModule } from '../report-nodes/report-nodes.module';
 import { ExecutionTabsComponent } from './components/execution-tabs/execution-tabs.component';
 import './components/execution-tabs/execution-tabs.component';
 import { CustomCellRegistryService, EntityRegistry } from '@exense/step-core';
-import { PlanIconComponent } from '../plan/components/plan-icon/plan-icon.component';
-import { PlanLinkComponent } from '../plan/components/plan-link/plan-link.component';
+import { ExecutionItemsSubtabsComponent } from './components/execution-item-subtabs/execution-items-subtabs.component';
 
 @NgModule({
   declarations: [
@@ -28,9 +27,10 @@ import { PlanLinkComponent } from '../plan/components/plan-link/plan-link.compon
     PanelIdPipe,
     KeywordCallsComponent,
     ExecutionTabsComponent,
+    ExecutionItemsSubtabsComponent,
   ],
   imports: [StepCommonModule, OperationsModule, ReportNodesModule],
-  exports: [ExecutionListComponent, ExecutionStepComponent, ExecutionTabsComponent],
+  exports: [ExecutionListComponent, ExecutionStepComponent, ExecutionTabsComponent, ExecutionItemsSubtabsComponent],
 })
 export class ExecutionModule {
   constructor(_entityRegistry: EntityRegistry, _cellsRegister: CustomCellRegistryService) {
