+++ projects/step-frontend/src/lib/modules/execution/services/executions-tabs-control.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@angular/core';
+import { Injectable, OnDestroy } from '@angular/core';
 import { AJS_MODULE } from '@exense/step-core';
 import { downgradeInjectable, getAngularJSGlobal } from '@angular/upgrade/static';
 import { Subject } from 'rxjs';
@@ -17,19 +17,24 @@ export type ExecutionTabsCommand = {
 @Injectable({
   providedIn: 'root',
 })
-export class ExecutionsTabsControlService {
-  readonly controlSubject$: Subject<ExecutionTabsCommand> = new Subject<ExecutionTabsCommand>();
+export class ExecutionsTabsControlService implements OnDestroy {
+  private commandSubject$: Subject<ExecutionTabsCommand> = new Subject<ExecutionTabsCommand>();
+  readonly command$ = this.commandSubject$.asObservable();
 
   /**
    * disables first tab of the given id of executionId
    * @param label
    */
   disableTab(executionId: string, id: string): void {
-    this.controlSubject$.next({
+    this.commandSubject$.next({
       command: ExecutionTabsCommandType.DISABLE,
       executionId: executionId,
       id: id,
     });
+  }
+
+  ngOnDestroy(): void {
+    this.commandSubject$.complete();
   }
 }
